Working version of the translate with dynamic batching.
Took me a bit to figure out how to do the bookkeeping to know which document chunk (sentence) goes with which document and in what order.

First attempted tried to encode the request_id for the InferenceRequest to Seamless with that information. However, there
doesn't seem to be a way to get the request_id from the resulting InferenceResponse object that gets returned.